# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Point of Sales System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List all purchases
      parameters:
        - name: supplierId
          in: query
          schema:
            type: ["integer", "null"]
        - name: from
          in: query
          schema:
            type: ["string", "null"]
        - name: to
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseWithDetails"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create a purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseWithDetails"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseWithDetails"
        "404":
          description: Not found
    delete:
      operationId: deletePurchase
      tags: [purchases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Purchase Returns
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List all sales
      parameters:
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: from
          in: query
          schema:
            type: ["string", "null"]
        - name: to
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaleWithDetails"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleWithDetails"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleWithDetails"
        "404":
          description: Not found
    delete:
      operationId: deleteSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Sales Returns
...
components:
  schemas:
...
    PurchaseDetail:
      type: object
      properties:
        id:
          type: integer
        purchaseId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        cost:
          type: number
        price:
          type: number
        discount:
          type: number
        qty:
          type: number
      required:
        - id
        - purchaseId
        - productId
        - productName
        - productCode
        - cost
        - price
        - discount
        - qty
    Purchase:
      type: object
      properties:
        id:
          type: integer
        purchaseDate:
          type: string
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        supplierBillNo:
          type: ["string", "null"]
        totalAmount:
          type: number
        createdAt:
          type: string
      required:
        - id
        - purchaseDate
        - totalAmount
        - createdAt
    PurchaseWithDetails:
      allOf:
        - $ref: "#/components/schemas/Purchase"
        - type: object
          properties:
            details:
              type: array
              items:
                $ref: "#/components/schemas/PurchaseDetail"
          required:
            - details
    CreatePurchaseDetailItem:
      type: object
      properties:
        productId:
          type: integer
        cost:
          type: number
        price:
          type: number
        discount:
          type: number
        qty:
          type: number
      required:
        - productId
        - cost
        - price
        - qty
    CreatePurchaseBody:
      type: object
      properties:
        purchaseDate:
          type: string
        supplierId:
          type: ["integer", "null"]
        supplierBillNo:
          type: ["string", "null"]
        details:
          type: array
          items:
            $ref: "#/components/schemas/CreatePurchaseDetailItem"
      required:
        - purchaseDate
        - details
...
    SaleDetail:
      type: object
      properties:
        id:
          type: integer
        saleId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        salesPrice:
          type: number
        discount:
          type: number
        qty:
          type: number
      required:
        - id
        - saleId
        - productId
        - productName
        - productCode
        - salesPrice
        - discount
        - qty
    PaymentTransaction:
      type: object
      properties:
        id:
          type: integer
        saleId:
          type: integer
        paymentType:
          type: string
          enum: [cash, cheque, card, voucher, loyalty]
        amount:
          type: number
      required:
        - id
        - saleId
        - paymentType
        - amount
    Sale:
      type: object
      properties:
        id:
          type: integer
        saleDate:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        billNo:
          type: string
        totalAmount:
          type: number
        createdAt:
          type: string
      required:
        - id
        - saleDate
        - billNo
        - totalAmount
        - createdAt
    SaleWithDetails:
      allOf:
        - $ref: "#/components/schemas/Sale"
        - type: object
          properties:
            details:
              type: array
              items:
                $ref: "#/components/schemas/SaleDetail"
            payments:
              type: array
              items:
                $ref: "#/components/schemas/PaymentTransaction"
          required:
            - details
            - payments
    CreateSaleDetailItem:
      type: object
      properties:
        productId:
          type: integer
        salesPrice:
          type: number
        discount:
          type: number
        qty:
          type: number
      required:
        - productId
        - salesPrice
        - qty
    CreatePaymentItem:
      type: object
      properties:
        paymentType:
          type: string
          enum: [cash, cheque, card, voucher, loyalty]
        amount:
          type: number
      required:
        - paymentType
        - amount
    CreateSaleBody:
      type: object
      properties:
        saleDate:
          type: string
        customerId:
          type: ["integer", "null"]
        details:
          type: array
          items:
            $ref: "#/components/schemas/CreateSaleDetailItem"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/CreatePaymentItem"
      required:
        - saleDate
        - details
        - payments
...