# lib/api-zod/src/generated/types/createSaleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Point of Sales System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePaymentItem } from "./createPaymentItem";
import type { CreateSaleDetailItem } from "./createSaleDetailItem";

export interface CreateSaleBody {
  saleDate: string;
  /** @nullable */
  customerId?: number | null;
  details: CreateSaleDetailItem[];
  payments: CreatePaymentItem[];
}